api/requirements.txt
pynest-api==0.2.1
asyncpg==0.29.0
prometheus-client==0.19.0
opentelemetry-distro==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-exporter-otlp==1.22.0